ai-fake-detect-system/backend/requirements.txt
flask
flask-cors
pillow
opencv-python
numpy
jieba
